added userId for Meal and changed InMemoryMealRepositoryImpl

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -5,9 +5,11 @@ import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -15,6 +17,7 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     {
         MealsUtil.MEALS.forEach(this::save);
+
     }
 
     @Override
@@ -29,18 +32,29 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id, int userId) {
+        if (repository.get(id).getUserId() == userId) {
+            repository.remove(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public Meal get(int id) {
+    public Meal get(int id, int userId) {
+        if (repository.get(id).getUserId() != userId) {
+            return null;
+        }
         return repository.get(id);
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(int userId) {
+        repository.values().stream().forEach(m -> m.getUserId().equals(userId));
+        return repository.values().stream().filter(s -> s.getUserId().equals(userId)).
+                sorted(Comparator.comparing(Meal::getDateTime).reversed()).
+                collect(Collectors.toList());
     }
+
 }
